Resolve package splitting http-service
